Moved Font Awesome tag from index.js to index.html Updated Private Chat Paths in App.jsx Live Video Export Syntax Changes Updated PrivChat API
Priv Chat basic implimentation complete.
The priv chat component will establish a connection to the BE.
This component will keep an array of conversation state updated on App.jsx.
This state is distrobuted to Chat inbox and each individual chat message.
The Chat thread component renders all messages with a single corrospondent.
This component can call a POST api to create a new message.